requirements.txt
django==1.10.2
Pillow==3.3.1
codegen==1.0
coreapi==2.0.8
django-bootstrap3==7.0.1
django-bootstrap3-datetimepicker-2==2.4.2
django-cors-headers==1.1.0
django-filter==0.14.0
django-rest-swagger==2.0.6
django-solo==1.1.2
djangorestframework==3.4.5
drfdocs==0.0.11